Reduce asset size
Some of the values in 'ranges' file are guaranteed
to be small enough to fit into 2 bytes.

+++ library/src/main/java/com/maphon/russvy/RussvyAssetReader.java
@@ -113,10 +113,10 @@ public class RussvyAssetReader {
             }
             while(true) {
                 try {
-                    mManager.addRange(String.valueOf(os.readInt()),
+                    mManager.addRange(String.valueOf(os.readShort()),
                             String.valueOf(os.readInt()),
                             String.valueOf(os.readInt()),
-                            os.readInt(), os.readInt());
+                            os.readShort(), os.readShort());
                 } catch (IOException e) {
                     break;
                 }
